sit100/requirements.txt
amqp==5.2.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
boto3==1.36.25
botocore==1.36.25
cashflow==1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.0.6
django-decorator-include==3.0
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
django-user-agents==0.4.0
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.54.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
kombu==5.3.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.40
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.11.2
seaborn==0.13.2
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
whitenoise==6.8.2
wsproto==1.2.0
yarg==0.1.9
zipp==3.20.2
